Renamed class file
Renamed classes.py to DailyWACOutput.py
