+++ DailyWACOutput.py
